update RunProcess test
